move ThreadLocalVariable to utils
